requirements.txt
jdatetime==4.1.1
requests==2.31.0
python-dotenv==1.0.0